winegstreamer: Introduce a new wg_transform_create_mf helper.
(cherry picked from commit c1158337ab95462118aef966204597d908d30e59)

CW-Bug-Id: #20833
